sdx2_encode.c

+++ src/sdx2_encode.c
@@ -161,7 +161,7 @@ sdx2_encode_mono(const s16 *ibuf_,
     {
       curr_sample = ibuf_[i];
 
-      comp_sample = encode_sample((s32)curr_sample,(s32)prev_sample);
+      comp_sample = encode_sample(curr_sample,prev_sample);
 
       obuf_[i] = comp_sample;
